Add final HF export to olmo_core_finetune; share export_to_hf via olmo_core_utils
olmo_core_finetune.py previously saved nothing at end of training (checkpoints
only at checkpointing_steps intervals), so short runs produced no evaluable
model. Move dpo.py's export_to_hf into olmo_core_utils and call it after
trainer.fit() when the model came from an HF checkpoint. Give ab_sft_olmocore.sh
an explicit weka output_dir so the export lands in a discoverable location.
