+++ open_instruct/olmo_core_finetune.py
@@ -345,6 +345,14 @@ def main(args: SFTArguments, tc: dataset_transformation.TokenizerConfig) -> None
     trainer.fit()
     logger.info("Training complete.")
 
+    if use_hf_ckpt:
+        hf_model_path = os.path.join(args.checkpoint.output_dir, "hf_model")
+        olmo_core_utils.export_to_hf(
+            train_module.model, tc.tokenizer, hf_model_path, args.model.model_name_or_path, is_main_process
+        )
+    else:
+        logger.warning("Skipping final HF export: model was not loaded from an HF checkpoint.")
+
     teardown_training_environment()
